refactor: 불필요 useEffect 삭제 및 컴포넌트 분리
