+++ src/views/Search/components/Result/SearchResult.tsx
@@ -1,10 +1,13 @@
 import { css } from '@emotion/react';
-import { MutableRefObject, useRef } from 'react';
+import { MutableRefObject, useRef, useState } from 'react';
 
+import { getBarrierFreeInfo, getSearchKeyword } from '@/apis/public/search';
 import { BigInfoIcon } from '@/assets/icon';
 import Loading from '@/components/Loading';
+import PageLoading from '@/components/PageLoading';
 import PlaceCard from '@/components/PlaceCard';
 import { MAP_FACILITIES_API_KEY } from '@/constants/facilities';
+import { useInfiniteScroll } from '@/hooks/use-infinite-scroll';
 import { COLORS, FONTS } from '@/styles/constants';
 import { BarrierFreeItem, SearchItem } from '@/types/search';
 
@@ -12,15 +15,80 @@ import { getFilterList } from '../../constants/category';
 import { filterState } from '../../types/category';
 
 interface SearchResultProps {
-  placeData: (SearchItem & BarrierFreeItem)[];
-  targetElement: MutableRefObject<HTMLDivElement | null>;
-  loading: boolean;
+  searchWord: string;
   filterState: filterState;
   heartList: number[];
 }
 
 const SearchResult = (props: SearchResultProps) => {
-  const { placeData, targetElement, loading, filterState, heartList } = props;
+  const { searchWord, filterState, heartList } = props;
+
+  const [loading, setLoading] = useState(false);
+  const [placeData, setPlaceData] = useState<(SearchItem & BarrierFreeItem)[]>(
+    [],
+  );
+
+  // 무한스크롤
+  const options = {
+    root: null,
+    rootMargin: '0px',
+    threshold: 0,
+  };
+
+  const handleObserver = async (
+    observer: IntersectionObserver,
+    target: MutableRefObject<HTMLElement | null>,
+    page: MutableRefObject<number>,
+  ) => {
+    setLoading(true);
+    const pageNo = page.current;
+
+    try {
+      const items = await getSearchKeyword({
+        pageNo,
+        numOfRows: 50,
+        MobileOS: 'ETC',
+        keyword: searchWord || '',
+        contentTypeId: 12,
+      });
+
+      if (items === '') {
+        if (pageNo === 0) setPlaceData([]);
+        target.current && observer.unobserve(target.current);
+      } else {
+        const placeData: (SearchItem & BarrierFreeItem)[] = [];
+        const promises = items.item.map(({ contentid }) =>
+          getBarrierFreeInfo({
+            MobileOS: 'ETC',
+            contentId: Number(contentid),
+          }),
+        );
+        const promiseResult = await Promise.allSettled(promises);
+        promiseResult.forEach((result) => {
+          if (result.status === 'fulfilled' && result.value !== '') {
+            const item = result.value.item;
+            const targetPlace = items.item.find(
+              ({ contentid }) => contentid === item[0].contentid,
+            );
+            if (!targetPlace) return;
+            placeData.push({ ...targetPlace, ...result.value.item[0] });
+          }
+        });
+
+        setPlaceData((prev) => [...prev, ...placeData]);
+        page.current++;
+      }
+    } finally {
+      setLoading(false);
+    }
+  };
+
+  const targetElement = useInfiniteScroll({
+    options,
+    handleObserver,
+    deps: [searchWord],
+  });
+
   const placeListRef = useRef<HTMLUListElement>(null);
 
   const filterList = getFilterList(filterState);
@@ -32,6 +100,7 @@ const SearchResult = (props: SearchResultProps) => {
 
   return (
     <>
+      {loading && placeData.length === 0 && <PageLoading />}
       <ul css={containerCss(placeData.length)} ref={placeListRef}>
         {!loading && renderPlaceList.length === 0 ? (
           <div css={noResultContainerCss}>
