+++ src/views/Search/pages/SearchResultPage.tsx
@@ -1,16 +1,12 @@
 import { css } from '@emotion/react';
-import { MutableRefObject, useEffect, useState } from 'react';
-import { useLocation, useParams } from 'react-router-dom';
+import { useState } from 'react';
+import { useParams } from 'react-router-dom';
 
-import { getBarrierFreeInfo, getSearchKeyword } from '@/apis/public/search';
 import getUserData from '@/apis/supabase/getUserData';
 import { SearchSetIcon } from '@/assets/icon';
 import MenuBar from '@/components/MenuBar';
-import PageLoading from '@/components/PageLoading';
 import { useAsyncEffect } from '@/hooks/use-async-effect';
-import { useInfiniteScroll } from '@/hooks/use-infinite-scroll';
 import { COLORS, FONTS } from '@/styles/constants';
-import { BarrierFreeItem, SearchItem } from '@/types/search';
 import { UserDataResponse } from '@/types/userAPI';
 import { isGuideShown } from '@/utils/storageHideGuide';
 
@@ -27,13 +23,9 @@ import { STORAGE_KEY } from '../constants/localStorageKey';
 import { category, filterState } from '../types/category';
 
 const SearchResultPage = () => {
-  const { word: initialWord } = useParams();
+  const { word: searchWord } = useParams();
 
-  const { pathname } = useLocation();
   const [userData, setUserData] = useState<UserDataResponse | null>(null);
-  const [placeData, setPlaceData] = useState<(SearchItem & BarrierFreeItem)[]>(
-    [],
-  );
 
   const [filterState, setFilterState] =
     useState<filterState>(INITIAL_FILTER_STATE);
@@ -44,13 +36,6 @@ const SearchResultPage = () => {
   );
   const [isFilterOpen, setIsFilterOpen] = useState(false);
 
-  // state handling func
-  const [loading, setLoading] = useState(false);
-
-  useEffect(() => {
-    setPlaceData([]);
-  }, [pathname]);
-
   useAsyncEffect(async () => {
     const kakaoId = sessionStorage.getItem('kakao_id');
     if (!kakaoId) return;
@@ -59,67 +44,6 @@ const SearchResultPage = () => {
     setUserData(userData);
     setFilterState(createInitialFilterState(userData?.universal_type || []));
   }, []);
-
-  // 무한스크롤
-  const options = {
-    root: null,
-    rootMargin: '0px',
-    threshold: 0,
-  };
-
-  const handleObserver = async (
-    observer: IntersectionObserver,
-    target: MutableRefObject<HTMLElement | null>,
-    page: MutableRefObject<number>,
-  ) => {
-    setLoading(true);
-    const pageNo = page.current;
-
-    try {
-      const items = await getSearchKeyword({
-        pageNo,
-        numOfRows: 50,
-        MobileOS: 'ETC',
-        keyword: pathname.split('/')[2],
-        contentTypeId: 12,
-      });
-
-      if (items === '') {
-        if (pageNo === 0) setPlaceData([]);
-        target.current && observer.unobserve(target.current);
-      } else {
-        const placeData: (SearchItem & BarrierFreeItem)[] = [];
-        const promises = items.item.map(({ contentid }) =>
-          getBarrierFreeInfo({
-            MobileOS: 'ETC',
-            contentId: Number(contentid),
-          }),
-        );
-        const promiseResult = await Promise.allSettled(promises);
-        promiseResult.forEach((result) => {
-          if (result.status === 'fulfilled' && result.value !== '') {
-            const item = result.value.item;
-            const targetPlace = items.item.find(
-              ({ contentid }) => contentid === item[0].contentid,
-            );
-            if (!targetPlace) return;
-            placeData.push({ ...targetPlace, ...result.value.item[0] });
-          }
-        });
-
-        setPlaceData((prev) => [...prev, ...placeData]);
-        page.current++;
-      }
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  const targetElement = useInfiniteScroll({
-    options,
-    handleObserver,
-    deps: [pathname],
-  });
 
   // 검색 가이드
   const handleSetShowGuide = (value: boolean) => {
@@ -152,17 +76,15 @@ const SearchResultPage = () => {
 
   return (
     <>
-      {loading && placeData.length === 0 && <PageLoading />}
       <div css={containerCss}>
-        <SearchBarContainer initialWord={initialWord}>
+        <SearchBarContainer initialWord={searchWord}>
           <button type="button" css={buttonCss} onClick={openFilter}>
             <SearchSetIcon />
             {selectedCategory()}
           </button>
           <SearchResult
-            placeData={placeData}
-            targetElement={targetElement}
-            loading={loading}
+            key={searchWord}
+            searchWord={searchWord || ''}
             filterState={filterState}
             heartList={userData?.favorite_list || []}
           />
